requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
roboflow>=1.1.0
python-dotenv>=1.0.0
